Add files via upload
I created my second table for the individual part of Iteration 5. I then added hyperlinking to the movie title so that when you click the movie title on the UI, it directs from the first table to the second table, with that movie info. We could eventually merge the code, but I'm putting it as a separate branch since the tables aren't complete and I don't want to muck any progress you guys made during 7/1/2020 in class.